// resources/app/libs/api-client/gen/models/release.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Bånder api docs
 * Bånder.App is a sophisticated media server developed with the primary goal to deliver high performance and seamless experience for users.
Being built on top of Laravel framework and PostgreSQL for database, it leverages the power of robust backend technologies, providing high security, reliability, and scalability.

The main focus of Bånder.App is efficient media management and delivery.
This application is designed to handle large amounts of media files while ensuring quick and efficient access.
Implemented queuing functionality through Redis offers efficient job management and load handling.
The media information is organized and easily searchable, ensuring users can always find what they are looking for quickly and easily.
 * OpenAPI spec version: 0.0.1
 */

export interface Release {
  id?: number;
  title?: string;
  uri?: string;
  resource_url?: string;
  type?: string;
  thumbnail?: string;
  cover_image?: string;
  images?: unknown[];
  artists?: unknown[];
  formats?: unknown[];
  labels?: unknown[];
  tracklist?: unknown[];
  country?: string;
  released?: string;
  year?: number;
  notes?: string;
  data_quality?: string;
  genres?: unknown[];
  styles?: unknown[];
  master_id?: string;
  master_url?: string;
  catno?: string;
}